Correct opening system-out balise

+++ src/Data_Format/util/LET_util_format_string.c
@@ -90,7 +90,7 @@ void LET_uint_to_decimal_string(char str[], uint64_t num){
   for (i = 0; i < len; i++){
     rem = (uint8_t)(num % 10);
     num = num / 10;
-    *((uint8_t*)&str[len - (i + 1)]) = rem + '0';
+    *((uint8_t*)&str[len - (i + 1)]) = (uint8_t)(rem + '0');
   }
 
   str[len] = '\0';
@@ -126,12 +126,12 @@ void LET_int_to_string(char str[], int64_t num){
 }
 
 
-int8_t LET_compare_str(const char * a, const char * b){
+int16_t LET_compare_str(const char * a, const char * b){
   while( ( *a != '\0' && *b != '\0' ) && *a == *b ){
     a++;
     b++;
   }
-  return *a - *b;
+  return (int16_t)(*a - *b);
 }
 
 
@@ -141,12 +141,12 @@ void LET_array_convert(char dest[], const char *const src, uint32_t size, LET_AS
   uint64_t buffer = 0;
   for (; i<size/whitespace; i++){
     buffer = LET_str_to_uint64(src+(i*whitespace), whitespace);
-    position +=  LET_convert_uint_to_base(dest+position, buffer, format, whitespace);
+    position +=  (uint16_t)LET_convert_uint_to_base(dest+position, buffer, format, whitespace);
     dest[position++]=' ';
   }
   for(uint8_t j = 0; j < size%whitespace; j++){
     buffer = LET_str_to_uint64(src+j+(i*whitespace), LET_BYTE);
-    position += LET_convert_uint_to_base(dest+position, buffer, format, LET_BYTE);
+    position += (uint16_t)LET_convert_uint_to_base(dest+position, buffer, format, LET_BYTE);
   }
   dest[position-((size%whitespace)?0:1)]='\0';
 }
